+++ CHANGELOG.md
@@ -10,8 +10,8 @@
 
 ### Additions and Improvements
 
-- New CLI flag `--get-blobs-api-p2p-sidecars-download-enabled` allows the beacon node to serve `getBlobs` REST API responses by attempting to fetch missing blob sidecars from the p2p network. The new flag `--get-blobs-sidecars-download-timeout` controls the network fetch timeout (default: 5 seconds). 
-- Update Netty to 4.2.9.Final
+- New CLI flag `--rest-api-getblobs-sidecars-download-enabled` allows the beacon node to serve `getBlobs` REST API responses by attempting to fetch missing blob sidecars from the p2p network. The new flag `--rest-api-getblobs-sidecars-download-timeout` controls the network fetch timeout (default: 5 seconds). 
+
 
 ### Bug Fixes